+++ solver_rust_conic/src/lib.rs
@@ -57,7 +57,7 @@ This crate includes two [`linalg::LinAlgEx`] implementors:
 
 ```toml
 [dependencies.totsu]
-version = "0.8.0"
+version = "0.8.1"
 features = ["f64lapack"]
 ```
 
@@ -71,7 +71,7 @@ Use this in `Cargo.toml`:
 
 ```toml
 [dependencies.totsu]
-version = "0.8.0"
+version = "0.8.1"
 default-features = false
 features = ["libm"]
 ```