+++ solver_rust_conic/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Update Rust edition to 2021.
 ### Deprecated
 ### Removed
 ### Fixed
@@ -16,8 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.8.1] - 2021-08-17
 ### Changed
 - Bug fix of `ProbSDP`.
-- Update version of `cblas`
-- Update versions of `float_eq` and `env_logger`
+- Update version of `cblas`.
+- Update versions of `float_eq` and `env_logger`.
 
 ## [0.8.0] - 2021-05-30
 ### Changed